extend tests with withdrawals
